backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai-whisper
mistralai
chromadb
python-multipart
